refactor JWT implementation into helper
